+++ NEWS.md
@@ -1,8 +1,11 @@
-
 ## stylo news ##
 
+### version 0.6.7, 2018/05/12
+  * supports the following taggers: TaKIPI (for Polish), Alpino (Dutch)
+  * the Imposters method reimplemented, via the new function imposters()
 
-### version 0.6.6, 2018/02/14
+
+### version 0.6.6, 2018/04/13
   * Cosine Delta implemented and aviable via GUI
   * Min-Max distance implemented
   * Entropy distance implemented